0604 코드 수정

+++ database_project/src/database_project/deleteTable_owner.java
@@ -182,7 +182,7 @@ public class deleteTable_owner extends JFrame implements ActionListener {
 					}
 					
 					for(int i = 0; i < delete_owner_id.size(); i++) {
-						String deleteStmt = "DELETE FROM DB2022_SALE WHERE owner_id = ?";
+						String deleteStmt = "DELETE FROM DB2022_SALE WHERE owner_id IN (SELECT owner_id FROM DB2022_OWNER WHERE owner_id = ?)";
 						String deleteStmt2 = "DELETE FROM DB2022_OWNER WHERE owner_id = ?";
 						PreparedStatement p = conn.prepareStatement(deleteStmt);
 						PreparedStatement p2 = conn.prepareStatement(deleteStmt2);
